python/requirements.txt
websocket-client
pika